pyproject.toml
[tool.poetry]
name = "blood scan report"
version = "0.1.0"
description = ""
authors = ["Devyansh Vakharia <devyanshvakharia2510@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pdfplumber = "^0.11.5"
python-dotenv = "^1.0.1"
google-generativeai = "^0.8.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
fastapi = "^0.115.6"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
